Update consumer.pyx
switch to 64 bit uint.

+++ dtrace_cython/consumer.pyx
@@ -90,7 +90,7 @@ cdef int walk(const dtrace_aggdata_t * data, void * arg) with gil:
         address = data.dtada_data + rec.dtrd_offset
 
         # TODO: need to extend this.
-        if rec.dtrd_size == sizeof(uint32_t):
+        if rec.dtrd_size == sizeof(uint64_t):
             keys.append((<int32_t *>address)[0])
         if rec.dtrd_size == 20 * sizeof(uint64_t):
             # case stack() has been used --> need to lookup symbols.
